requirements.txt
# --- Computer Vision & Deep Learning ---
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
deep-sort-realtime>=1.3.2

# --- Data Engineering & Streaming ---
kafka-python>=2.0.2
pydantic>=2.0.0

# --- Graph Database & Data Handling ---
neo4j>=5.10.0
pandas>=2.0.0
numpy>=1.24.0

# --- Dashboard & Visualization ---
streamlit>=1.25.0
pyvis>=0.3.2
networkx>=3.1

# --- Utilities ---
python-dotenv>=1.0.0  # Check karo ki is line ke baad kuch extra toh nahi hai
PyYAML>=6.0
tqdm>=4.65.0